fix: exposeport upgrade
